Moved attributeToOptions to module
