+++ src/js/permutive.js
@@ -1,39 +1,11 @@
+import merge from 'lodash.merge';
 import bootstrap from './bootstrap';
 import api from './api';
-import merge from 'lodash.merge';
-const ATTRIBUTE_PATTERN = 'oPermutive';
+import identifyUser from './identifyUser';
+import { attributeToOption } from './attributes'
 
-const OPTION_PARENT_NODES = [
-	'adsApi',
-	'appInfo',
-	'contentApi',
-	'contentId',
-	'oComponent',
-	'publicApiKeys',
-	'userApi'
-];
-
-const formatOptionName = key => {
-	const keyLower = key.toLowerCase();
-
-	for (const mapped of OPTION_PARENT_NODES) {
-		if (mapped.toLowerCase() === keyLower) {
-			return mapped;
-		}
-	}
-
-	return keyLower;
-};
-
-const validateOptions = ({ opts, oPermutiveEl }) => {
-	const options = Object.assign({}, opts || Permutive.getDataAttributes(oPermutiveEl));
-
-	if (!options.publicApiKeys) {
-		throw new Error('o-permutive: No public API Keys found in options.');
-	}
-
-	return options;
-};
+const getPScriptURI = permutiveApiId =>
+	`https://cdn.permutive.com/${permutiveApiId}-web.js`;
 
 // TODO Consents can be derived outside of the package and passed in as config.
 function getConsents() {
@@ -53,18 +25,30 @@ function getConsents() {
 }
 
 function attachPermutiveScript(options) {
-	const url = "https://cdn.permutive.com/" + options.publicApiKeys.id + "-web.js";
-	if(!document.querySelector(`script[src="${url}"]`)) {
-		const s = document.createElement("script");
+	const permutiveURI = getPScriptURI(options.publicApiKeys.id)
+
+	if (!document.querySelector(`script[src="${permutiveURI}"]`)) {
+		const scriptTag = document.createElement("script", {
+			async: "true",
+			type: "text/javascript",
+			id: "permutive-script",
+			src: permutiveURI,
+		});
+
 		const HEAD = document.head || document.getElementsByTagName('head')[0];
-		s.async = "true";
-		s.type = "text/javascript";
-		s.id = "permutive-script";
-		s.src = "https://cdn.permutive.com/" + options.publicApiKeys.id + "-web.js";
-		HEAD.appendChild(s);
+		HEAD.appendChild(scriptTag);
 	}
 }
 
+const validateOptions = ({ opts, oPermutiveEl }) => {
+	const options = Object.assign({}, opts || Permutive.getDataAttributes(oPermutiveEl));
+
+	if (!options.publicApiKeys) {
+		throw new Error('o-permutive: No public API Keys found in options.');
+	}
+
+	return options;
+};
 
 class Permutive {
 	/**
@@ -102,32 +86,6 @@ class Permutive {
 	}
 
 	/**
-	 * Extract the option's path from an attribute name in camelCase form - coming from the component's dataset -
-	 * and returns a single element object.
-	 * @param {String} optKey - The attribute name in camelCase form, taken from the component's dataset. e.g., publicapikeysId
-	 * @param {String} optValue - The value assigned to optKey.
-	 * @returns {Object} - An object containing a single { key: "value" } or { key: { subkey: value } }
-	 */
-	static attributeToOption({ optKey, optValue }) {
-		const regex = new RegExp(`(^${ATTRIBUTE_PATTERN})?([A-Z][a-z]+)`, 'g');
-		const [/* mWhole */, mPrefix, mOpt] = regex.exec(optKey) || [];
-
-		const shortOptKey = mPrefix
-			? mOpt
-				? mOpt
-				: optKey
-			: optKey;
-
-		const [/* mWhole2 */, /* mPrefix2 */, mOpt2] = regex.exec(optKey) || [];
-
-		return {
-			[formatOptionName(shortOptKey)]: mOpt2
-				? { [formatOptionName(mOpt2)]: optValue }
-				: optValue,
-		};
-	}
-
-	/**
 	 * Get the data attributes from the PermutiveElement. If the component is being set up
 	 * declaratively, this method is used to extract the data attributes from the DOM.
 	 * @param {HTMLElement} oPermutiveEl - The component element in the DOM
@@ -139,7 +97,7 @@ class Permutive {
 		}
 
 		return merge({}, ...Object.keys(oPermutiveEl.dataset)
-			.map((optKey) => this.attributeToOption({ optKey, optValue: oPermutiveEl.dataset[optKey] }))
+			.map((optKey) => attributeToOption({ optKey, optValue: oPermutiveEl.dataset[optKey] }))
 		);
 	}
 
@@ -174,10 +132,8 @@ class Permutive {
 	 * @param {Object} userDemog
 	 * @param {Object} pageMeta
 	 */
-	static pAddon(userDemog, pageMeta) {
-		let user = { "user": Object.assign(userDemog) };
-		let data = { "page": Object.assign(pageMeta, user) };
-		window.permutive.addon('web', data);
+	static pAddon(page = { type, user, page }) {
+		window.permutive.addon('web', page);
 	}
 
 	/**
